# Conflicts:
#	GameJam/.vs/GameJam/v15/Server/sqlite3/storage.ide
#	GameJam/Library/CurrentLayout.dwlt
#	GameJam/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	GameJam/Library/assetDatabase3
#	GameJam/Library/expandedItems

+++ GameJam/Assets/Scripts/ConfigurationObject.cs
@@ -8,17 +8,20 @@ using UnityEngine;
 public class ConfigurationObject : ScriptableObject {
 	public float startPlayerCredits = 100.0f;
 	public float maxPlayerCredits = 200.0f;
-	public float BarracksHP = 30.0f;
-	public float FarmHP = 10.0f;
-	[Header("BaseBuy")]
+    public int BarracksHP = 30;
+    public int FarmHP = 10;
+
+    [Header("BaseBuy")]
 	public float BarracksBuyCost = 20.0f;
 	public float FarmBuyCost = 20.0f;
-	[Header("BaseOperate")]
+
+    [Header("BaseOperate")]
 	public float FarmEarn = 10.0f;
 	public float FarmEarnPeriod = 5.0f;
 	public float CostForSoldier = 5.0f;
 	public float BarrackSpawnEverySec = 5.0f;
-	[Header("Other")]
+
+    [Header("Other")]
 	public float InfoTextDisplayTime = 3.0f;
     public float AIDecisionEverySec = 2.1f;
     //public float CameraMinZoom = 0.2f;
